backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
pytest==8.0.0
pytest-django==4.8.0
coverage==7.4.1
setuptools<70.0.0
dj-database-url>=2.1.0
setuptools<70.0.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
pillow>=10.2.0